Move data ref to it's own folder
